Create module for switch the active state of the page
