+++ js/form.js
@@ -6,7 +6,6 @@ import {addLoadPhotoListeners, removeLoadPhotoListeners} from './avatar.js';
 const URL_SEND = 'https://23.javascript.pages.academy/keksobooking';
 
 const adForm = document.querySelector('.ad-form');
-const adFilters = document.querySelector('.map__filters');
 const adPriceInput = adForm.querySelector('#price');
 const adRoomsSelect = adForm.querySelector('#room_number');
 const adGuestsSelect = adForm.querySelector('#capacity');
@@ -166,27 +165,4 @@ const removeFormListeners = () => {
   removeFiltersSelectListener(filtersCallback);
 };
 
-//Activation/Disactivation of form
-const changeDisabledStatusOfElementWithChildren = (elementContainer, className, isDisabled, actionName) => {
-  elementContainer.classList[actionName](className);
-  for (const element of elementContainer.children) {
-    element.disabled = isDisabled;
-  }
-};
-
-const switchToInactiveState = () => {
-  changeDisabledStatusOfElementWithChildren(adForm, 'ad-form--disabled', true, 'add');
-  changeDisabledStatusOfElementWithChildren(adFilters, 'map__filters--disabled', true, 'add');
-  removeFormListeners();
-};
-
-const switchFormToActiveState = () => {
-  changeDisabledStatusOfElementWithChildren(adForm, 'ad-form--disabled', false, 'remove');
-  addFormListeners();
-};
-
-const switchFiltersToActiveState = () => {
-  changeDisabledStatusOfElementWithChildren(adFilters, 'map__filters--disabled', false, 'remove');
-};
-
-export {switchFormToActiveState, switchFiltersToActiveState, switchToInactiveState};
+export {addFormListeners, removeFormListeners};
